sorry for doing this on main lol. Added inv quantity check in order checkout, decrement seller quant, increment seller balance, connected placing an order to orderfact and order contents tables

+++ app/cart.py
@@ -6,6 +6,7 @@ from wtforms import StringField, PasswordField, BooleanField, SubmitField
 from wtforms.validators import ValidationError, DataRequired, Email, EqualTo
 from flask import jsonify
 from decimal import Decimal
+import datetime
 
 
 from .models.user import User
@@ -14,6 +15,7 @@ from .models.carts import Cart, CartContents
 from .models.orders import Order, OrderItem
 from .models.balance import Balance
 from .models.ordercontents import OrderHistory
+from .models.inventory import Inventory
 
 
 
@@ -123,21 +125,61 @@ def place_order():
         total_cost = 0
         total_products = 0
 
+
+
+
+
+
+
+
+
+
+
+
+
+
     if balance > total_cost:
         order_info = CartContents.get_cart(user_id)
+        
         if cart_info:
-            flash('Order placed successfully!', 'success')
-            #OrderHistory.insert_user_purchase_history(user_id, order_info)
-            for item in cart_info:
-                CartContents.delete_from_cart(user_id, item.product_id, item.seller_id)
-            #balance = float(balance) - float(total_cost)
-            '''amount = Decimal(total_cost)
-            new_balance = Balance.calculate_new_balance(user_id, -amount)
-            Balance.insert_new_balance(user_id, new_balance)'''
-            return render_template('orders.html', orders=order_info, total_cost=total_cost, total_products=total_products)
+
+            if CartContents.check_inventory(user_id):
+            
+                flash('Order placed successfully!', 'success')
+                #OrderHistory.insert_user_purchase_history(user_id, order_info)
+                CartContents.add_order_to_orderfact(current_user.id, total_cost, time_purchased=datetime.datetime.now(), fulfill_status=False)
+                    
+                for item in cart_info:
+                    
+                    CartContents.add_order_to_ordercontents(CartContents.get_max_orderfact_id(), item.product_id, item.seller_id, item.quantity)
+                    CartContents.increment_seller_balance(user_id=item.seller_id, total_price= (item.quantity * item.price))
+                    Inventory.update_inventory_quantity(pid=item.product_id, uid=item.seller_id, quant = Inventory.get_current_quantity(item.seller_id, item.product_id) - item.quantity) # need to get a way to get original quantity
+                    CartContents.delete_from_cart(user_id, item.product_id, item.seller_id)
+                #balance = float(balance) - float(total_cost)
+                '''amount = Decimal(total_cost)
+                new_balance = Balance.calculate_new_balance(user_id, -amount)
+                Balance.insert_new_balance(user_id, new_balance)'''
+                return render_template('orders.html', orders=order_info, total_cost=total_cost, total_products=total_products)
+            else:
+                flash('You are placing an order for more than the seller has in stock for at least one item. Decrease your quantity and try again.', 'danger')
+            return redirect(url_for('cart.cart'))
+     
         else:
             flash('Failed to place the order. Please try again later.', 'danger')
             return redirect(url_for('cart.cart'))
+     
+     
+     
+     
+     
+     
+     
+     
+     
+     
+        
+        
+        
     else:
         if balance < total_cost:
             flash('Insufficient funds. Please make a deposit and try again.', 'danger')
